requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.10
langchain-groq==0.0.1
langchain-google-genai==0.0.6
chromadb==0.4.22
pandas==2.1.4
pydantic==2.5.2
numpy==1.24.3
